airflow/requirements-airflow.txt
duckdb==0.9.2
apache-airflow-providers-docker==4.2.1
python-dotenv